+++ examples/ex_ppo.py
@@ -22,7 +22,7 @@ env=CreateEnvironment(method='ppo', fit=Sphere,
                       bounds=bounds, mode='min', episode_length=50)
 
 #create a callback function to log data
-cb=RLLogger(check_freq=1, mode='min')
+cb=RLLogger(check_freq=1)
 #create a RL object based on the env object
 ppo = PPO2(MlpPolicy, env=env, n_steps=12, seed=1)
 #optimise the enviroment class