+++ neorl/tests/test_dqn.py
@@ -29,7 +29,7 @@ def test_dqn():
                           mode='min', 
                           episode_length=50)
     #create a callback function to log data
-    cb=RLLogger(check_freq=1, mode='min')
+    cb=RLLogger(check_freq=1)
     #create an a2c object based on the env object
     dqn = DQN(DQNPolicy, env=env)
     #optimise the enviroment class