+++ neorl/tests/test_ppo2.py
@@ -25,7 +25,7 @@ def test_ppo():
     env=CreateEnvironment(method='ppo', fit=Sphere, 
                           bounds=bounds, mode='min', episode_length=50)
     #create a callback function to log data
-    cb=RLLogger(check_freq=1, mode='min')
+    cb=RLLogger(check_freq=1)
     #create an a2c object based on the env object
     ppo = PPO2(MlpPolicy, env=env, n_steps=12)
     #optimise the enviroment class